+++ app/src/main/java/org/cazait/cazait_android/data/model/remote/request/MenuRequest.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class MenuRequest(
     @SerializedName("cafeId")
-    val cafeId: Int
+    val cafeId: Long
 )
