+++ app/src/main/java/org/cazait/cazait_android/data/repository/DataRepository.kt
@@ -10,5 +10,5 @@ interface DataRepository {
     suspend fun addToFavourite(cafeId: String): Flow<Resource<Boolean>>
     suspend fun removeFromFavourite(cafeId: String): Flow<Resource<Boolean>>
     suspend fun getCafes(userId: Long, query: CafeListRequest): Flow<Resource<CafeListResponse>>
-    suspend fun getMenus(cafeId:Int):Flow<Resource<MenuResponse>>
+    suspend fun getMenus(cafeId:Long):Flow<Resource<MenuResponse>>
 }
